multi_dims/requirements.txt
torch==2.3.0
torchvision>=0.18.0
vllm==0.7.1
transformers==4.41.0
accelerate>=0.28.0
sentence-transformers>=2.7.0
openai==1.30.1
datasets>=2.19.0
scikit-learn>=1.4.0
scipy>=1.12.0
numpy>=1.26.0
pydantic>=2.7.0,<3.0
typing_extensions>=4.7.0
unidecode>=1.3.8
tqdm>=4.66.0
sentencepiece>=0.1.99
python-dotenv>=0.19.0
networkx>=3.0
matplotlib>=3.5.0